Borro herencia de ESQuery
    refs:#189
